separate tests for common backend and matplotlib backend
