+++ nilearn/plotting/surface/tests/test_backend.py
@@ -1,0 +1,36 @@
+import pytest
+
+from nilearn.plotting.surface._backend import (
+    _check_hemisphere_is_valid,
+    _check_view_is_valid,
+)
+
+
+@pytest.mark.parametrize(
+    "view,is_valid",
+    [
+        ("lateral", True),
+        ("medial", True),
+        ("latreal", False),
+        ((100, 100), True),
+        ([100.0, 100.0], True),
+        ((100, 100, 1), False),
+        (("lateral", "medial"), False),
+        ([100, "bar"], False),
+    ],
+)
+def test_check_view_is_valid(view, is_valid):
+    assert _check_view_is_valid(view) is is_valid
+
+
+@pytest.mark.parametrize(
+    "hemi,is_valid",
+    [
+        ("left", True),
+        ("right", True),
+        ("both", True),
+        ("lft", False),
+    ],
+)
+def test_check_hemisphere_is_valid(hemi, is_valid):
+    assert _check_hemisphere_is_valid(hemi) is is_valid
